Proposal for GSoC project about ROOT-git integrantion
This is my personal idea for a ROOT related GSoC. It should be self
contained enough that any good CS student without a HEP background
should be able to have results in a couple of months.

Moreover if works out as expected this could provide a very simple
yet elegant solution for a number of usecases.

I am looking for a co-mentor from the ROOT-side.