Change docker invocation command
To address some issues with using a docker container with an unknown UID
internally (user=$(id -u)) it's better to run with an explicit HOME set to /tmp.
This allows the docker container to run properly for both website preview and
html proofing, without giving spurious warnings.

I have an updated container ready to tag, this change adapts to it correctly.