Fix meeting navigation and move calendar (#383)
* Fix meeting navigation and move calendar

Meeting navigation implemented in 3712e72ef didn't work in practice as
GitHub pages doesn't allow Jeykll plugins. This rework uses the builtin
page.next/previous navigation instead. This has the drawback that
all posts (including announcements) are part of the navigation,
instead of just meeting. However, it's a substantial improvement
on having no navigation at all.

The community calendar was moved from the "activities" dropdown
to "communication", as discussed and approved in the meeting
of 2018-09-20.

* Remove URL checking from IOP Science site: It's too slow and frequently causes false positives