Add navigation links to event and newsletter layouts
As the events and newsletter templates are used in the general
list of navigated posts it's important to maintain the
forward and backward links on them so that users don't lose
navigation on these pages.

The navigation code is factorised out to an "include" and then
pulled in to the correct places of the event, meeting and
newsletter layouts.