Add minutes from last meeting and reorganise site for new WGs
Minutes from this week's meeting are added.

As the mandate for the new WGs have been agreed, this has
been converted from the Google Doc and added to the site.

Stubs for each of the new working groups have been created.
As listing all of these as "activities" made the menu item too large,
some current activities were also moved into the new
"workinggropus" section; the navbar builder now splits between
activities and working groups.

An announcement for the upcoming generators workshop was
added, as well as an events listing.