Fix gsoc_project.ext include
This include shows a header including either page.title if defined
or page.project. The current condition page.title == blank is never True
even if a proposal has not defined that attribute. This variable seems
to be set by default to "Project_<PROJECTNAME>". This PR modifies the
condition to print the content of page.title if "Project_" is not part
of page.title else page.project is shown.