Add meetings menu item and description
We now run a few series of meetings that are neither
working groups per-se or activities. It seemed then best
to add these are a separate set of topics, using the same
Jekyll engine way to process them automatically.

A new dropdown menu link is then added to the navbar
for these items.