Small fix to GSoC proposal title
The title is used in the full list of proposals https://hepsoftwarefoundation.org/gsoc/2024/summary.html, which misses the context of the project, so the project name needs to be repeated here.

This is just an intermediate fix to make the list presentable, we should very likely prefix the project name in the template for this list.