gsoc: Prepend project name in full list of proposals
The metadata format of the proposals

```
---
project: Ganga
title: Incorporate a Large Language Model to assist users
layout: gsoc_proposal
year: 2024
...
```

nudges authors to write the title in context of the project. In the full list of project proposals (which we have to submit to gsoc), this context is lost, so the project name has to be repeated in the title for it to make sense, in many cases.

I propose instead to fix the full list, prefixing the project name there.

This would mess with the order of the list though, so I update the sorting as well.
I couldn't get `sort: 'project'` to work (jekyll complains about not being able to do a comparison), but `sort: 'url'` is a nice proxy for project name + title.