gsoc: fix whitespaces in project/org summaries (#1530)
Due to the project/org descriptions being multiline yaml strings that had their newlines stripped, it's easy to mess up the whitespaces. This should bring an end to these woes, even if it makes the template a bit more roundabout (implementation from https://stackoverflow.com/questions/45147948/how-to-strip-newlines-in-jekyll). In my test, existing descriptions look OK, the whitespace in the Ganga one has been fixed, and it should no longer matter if you put newlines in `description: |`