fix: conveners list on HSF Training (#1537)
I was about to give up on #1512 and replace the fancy list with a simple bulleted list, but then I noticed that Lera's and Holly's names were not exactly the same as the `title` attribute in their profiles. So I'll try that first, and if it renders correctly, #1512 is fixed. If not, we use a bulleted list.