Improve project pages
Update the page of affiliated software with properly affiliated
projects, keeping the older umbrella projects for now (a redirect from
the old page is in place).

Remove the anachronistic "services" page.

Update menu to link to new pages (including a more obvious link to
the badges).

Update the project criteria to not refer to contracts for developers, but
rather to their commitment to provide support.
Reduce the Bronze level commitment to 1 year (from 2).